pushing

+++ src/main/java/com/johnson/app/rest/Repository/UserRepo.java
@@ -3,5 +3,5 @@ package com.johnson.app.rest.Repository;
 import com.johnson.app.rest.model.UserModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepo extends JpaRepository<UserModel,Long> {
+public interface UserRepo extends JpaRepository<UserModel,Integer> {
 }
